test-requirements.txt
mock>=1.0.1
pytest>=1.4.25
pytest-twisted
pytest-tornado